National Survey of Digital Game Players
The report of the most important information on the consumption of game media in Iran, 2021, has been published based on a national survey of digital game players. The Digital Games Research Center (DIREC) has been in charge of managing the design of questions, determining the sample size, survey methodology and monitoring the process. I have collaborated in this national project as supervisor and statistical analysis expert.